+++ src/apps/chifra/pkg/monitor/monitor_read.go
@@ -31,7 +31,7 @@ func (mon *Monitor) ReadMonitorHeader() (err error) {
 }
 
 // ReadAppearanceAt returns the appearance at the one-based index. The file remains open.
-func (mon *Monitor) ReadAppearanceAt(idx uint32, app *index.AppearanceRecord) (err error) {
+func (mon *Monitor) ReadAppearanceAt(idx int64, app *index.AppearanceRecord) (err error) {
 	if idx == 0 || idx > mon.Count() {
 		// the file contains a header one record wide, so a one-based index eases caller code
 		err = fmt.Errorf("index out of range in ReadAppearanceAt[%d]", idx)
@@ -64,7 +64,7 @@ func (mon *Monitor) ReadAppearanceAt(idx uint32, app *index.AppearanceRecord) (e
 // ReadAppearances returns appearances starting at the first appearance in the file. The call
 // will read as many records as are available in the array. The file remains opened.
 func (mon *Monitor) ReadAppearances(apps *[]index.AppearanceRecord) (err error) {
-	if uint32(len(*apps)) > mon.Count() {
+	if int64(len(*apps)) > mon.Count() {
 		err = fmt.Errorf("array is larger than the size of the file in ReadAppearances (%d,%d)", len(*apps), mon.Count())
 		return
 	}
